feat: add genre page, navigation and favorite support
Ajout de la page Genre avec son ViewModel, services de chargement et édition.
Navigation vers la page d’un genre depuis les pages Artiste, Track et Listening.
Affichage des albums, statistiques et image du genre sur la nouvelle page.
Ajout de la gestion du favori pour les genres (commandes et backend).
Affichage du genre dans les listes de morceaux, avec navigation possible.
Ajout de styles et ressources pour l’affichage du genre.
Corrections de libellés et traductions ("compilations", "Genre").
Refactorisation des types d’identifiants de genre (int → long).
Intégration complète dans la DI et le service de navigation.
Amélioration de l’expérience utilisateur autour des genres musicaux.

+++ Core/Rok.Application/Features/Tracks/Query/GetTracksByGenreIdQueryHandler.cs
@@ -2,10 +2,10 @@
 
 namespace Rok.Application.Features.Tracks.Query;
 
-public class GetTracksByGenreIdQuery(int genreId) : IQuery<IEnumerable<TrackDto>>
+public class GetTracksByGenreIdQuery(long genreId) : IQuery<IEnumerable<TrackDto>>
 {
     [RequiredGreaterThanZero]
-    public int GenreId { get; } = genreId;
+    public long GenreId { get; } = genreId;
 }
